refactoring kernel functions
