+++ kernel_embedding_dictionary/kernels/__init__.py
@@ -5,7 +5,7 @@
 from .expquad_kernel import ExpQuadKernel, ExpQuadKernelUni
 from .kernel import ProductKernel, UnivariateKernel
 from .matern12_kernel import Matern12Kernel, Matern12KernelUni
-from .matern13_kernel import Matern32Kernel, Matern32KernelUni
+from .matern32_kernel import Matern32Kernel, Matern32KernelUni
 from .matern52_kernel import Matern52Kernel, Matern52KernelUni
 from .matern72_kernel import Matern72Kernel, Matern72KernelUni
 from .matern_kernel import MaternNu2Kernel, MaternNu2KernelUni
