+++ kernel_embedding_dictionary/kernels/expquad_kernel.py
@@ -4,10 +4,8 @@
 
 from typing import List, Optional
 
-import numpy as np
-
-from ..utils import scaled_diff
 from .kernel import ProductKernel, UnivariateKernel
+from .kernel_funcs_1d import expquad_kernel_func_1d
 
 
 class ExpQuadKernelUni(UnivariateKernel):
@@ -18,18 +16,14 @@ class ExpQuadKernelUni(UnivariateKernel):
 
         self.ell = ell
 
-    def get_param_dict(self) -> dict:
+        self._kernel_func = expquad_kernel_func_1d
+
+    @property
+    def param_dict(self) -> dict:
         return {"ell": self.ell}
 
-    def _evaluate(self, x1: np.ndarray, x2: np.ndarray) -> np.ndarray:
-        n1 = x1.shape[0]
-        n2 = x2.shape[0]
-        K = np.zeros([n1, n2])
-        for i in range(n1):
-            for j in range(n2):
-                diff = 0.5 * scaled_diff(x1[i], x2[j], self.ell, np.sqrt(2))
-                K[i, j] = np.exp(-(diff**2))
-        return K
+    def _evaluate_pair(self, x1: float, x2: float) -> float:
+        return self._kernel_func(x1, x2, **self.param_dict)
 
 
 class ExpQuadKernel(ProductKernel):
