+++ kernel_embedding_dictionary/kernels/kernel.py
@@ -3,21 +3,36 @@
 
 
 import abc
-from typing import List
+from typing import List, Union
 
 import numpy as np
 
 
 class UnivariateKernel(abc.ABC):
+
+    @property
     @abc.abstractmethod
-    def get_param_dict(self) -> dict:
+    # TODO: turn into absract property
+    def param_dict(self) -> dict:
         pass
 
     @abc.abstractmethod
-    def _evaluate(self, x1: np.ndarray, x2: np.ndarray) -> np.ndarray:
+    def _evaluate_pair(self, x1: float, x2: float) -> float:
+        """use pass in case _evaluate is overridden"""
         pass
+
+    def _evaluate(self, x1: np.ndarray, x2: np.ndarray) -> np.ndarray:
+        """override for efficiency. x1 and x2 have shape (n1, ) and (n2, )."""
+        n1 = x1.shape[0]
+        n2 = x2.shape[0]
+        K = np.zeros([n1, n2])
+        for i in range(n1):
+            for j in range(n2):
+                K[i, j] = self._evaluate_pair(x1[i], x2[j])
+        return K
 
     def evaluate(self, x1: np.ndarray, x2: np.ndarray) -> np.ndarray:
+        """x1 and x2 have shape (n1, ) and (n2, )."""
         if (len(x1.shape)) != 1 or (len(x2.shape) != 1):
             raise ValueError(f"x1 ({x1.shape}) and x2 ({x2.shape}) must be one-dimensional arrays.")
 
@@ -42,7 +57,7 @@ class ProductKernel(abc.ABC):
         pass
 
     def get_param_dict_from_dim(self, dim: int) -> dict:
-        return self._kernels[dim].get_param_dict()
+        return self._kernels[dim].param_dict
 
     def evaluate(self, x1: np.ndarray, x2: np.ndarray) -> np.ndarray:
         if (len(x1.shape) != 2) or (len(x2.shape) != 2):
