+++ tests/kernel_embedding_dictionary/kernels/test_expquad_kernel.py
@@ -19,7 +19,7 @@ def test_expquad_kernel_uni_param_dict():
 
     ell = 1.5
     k = ExpQuadKernelUni(ell)
-    p = k.get_param_dict()
+    p = k.param_dict
 
     # this check is important due to how the kernel embedding params are assembled
     assert set(p.keys()) == {"ell"}
