+++ tests/kernel_embedding_dictionary/kernels/test_matern52_kernel.py
@@ -20,7 +20,7 @@ def test_matern52_kernel_uni_param_dict():
 
     ell = 1.5
     k = Matern52KernelUni(ell)
-    p = k.get_param_dict()
+    p = k.param_dict
 
     # this check is important due to how the kernel embedding params are assembled
     assert set(p.keys()) == {"ell", "nu"}
